requirements.txt
openai>=1.0.0
anthropic>=0.30.0
pyyaml>=6.0
tqdm>=4.60.0
